Make the ability to connect to tracks an interface
This also fixes a bug that was caused by using the wrong placement
method. Before we used `onBlockPlacedBy` which is called when the block
is already placed in the world, thus leading to issues where other
blocks thought the newly placed block already has connections. Now with
`getStateForPlacement` we can circumvent this issue.